Adds new NewsFeedPost class
